Cut out EnumerableExtensions to another file
